Renaming interfaces fileNames to comply with internal good practices
