+++ src/main/java/br/com/maykmoreira/apiexample/controller/MovieController.java
@@ -3,7 +3,7 @@ package br.com.maykmoreira.apiexample.controller;
 import br.com.maykmoreira.apiexample.customException.MovieNotFoundException;
 import br.com.maykmoreira.apiexample.model.Movie;
 
-import br.com.maykmoreira.apiexample.service.MovieService;
+import br.com.maykmoreira.apiexample.service.IMovieService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,23 +15,23 @@ import javax.validation.Valid;
 @RestController
 public class MovieController {
 
-    private MovieService movieService;
+    private IMovieService IMovieService;
 
     @Autowired
-    public MovieController(MovieService movieService) {
-        this.movieService = movieService;
+    public MovieController(IMovieService IMovieService) {
+        this.IMovieService = IMovieService;
     }
 
     @RequestMapping(value = "/movie", method = RequestMethod.GET)
     public List<Movie> get(){
-        return movieService.getAllMovies();
+        return IMovieService.getAllMovies();
     }
 
     @RequestMapping(value = "/movie/{id}", method = RequestMethod.GET)
     public ResponseEntity<Movie> getById(@PathVariable(value = "id") long id) {
 
         try{
-            return movieService.getMovieById(id);
+            return IMovieService.getMovieById(id);
         }catch (MovieNotFoundException e){
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
@@ -41,13 +41,13 @@ public class MovieController {
     @RequestMapping(value = "/movie", method =  RequestMethod.POST)
     public Movie post(@Valid @RequestBody Movie movie)
     {
-        return movieService.createMovie(movie);
+        return IMovieService.createMovie(movie);
     }
 
     @RequestMapping(value = "/movie/{id}", method = RequestMethod.DELETE)
     public ResponseEntity<Object> delete(@PathVariable(value = "id") long id)
     {
-        return movieService.deleteMovie(id);
+        return IMovieService.deleteMovie(id);
     }
 
 }
