+++ src/main/java/br/com/maykmoreira/apiexample/service/IMovieService.java
@@ -1,0 +1,14 @@
+package br.com.maykmoreira.apiexample.service;
+
+import br.com.maykmoreira.apiexample.customException.MovieNotFoundException;
+import br.com.maykmoreira.apiexample.model.Movie;
+import org.springframework.http.ResponseEntity;
+
+import java.util.List;
+
+public interface IMovieService {
+    public List<Movie> getAllMovies();
+    public ResponseEntity<Movie> getMovieById(Long id) throws MovieNotFoundException;
+    public Movie createMovie(Movie movie);
+    public ResponseEntity<Object> deleteMovie(Long id);
+}
